+++ readenv/loads.py
@@ -18,6 +18,6 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 
-from .readenv import loads
+from .readenv import load
 
-loads()
+load()