+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * fix typing support for python 3.8
 * made `readenv.redenv.load` "private" (`readenv.readenv._load`)
+* rename `readenv.readenv.loads` to `readenv.readenv.load`
 
 ## 0.1.1